created folders, updated namespaces
