+++ RTGS.TechTest.Api/Models/MyBalance.cs
@@ -1,0 +1,3 @@
+﻿namespace RTGS.TechTest.Api.Models;
+
+public record MyBalance(float Balance = 0);
